Experience (XP) should be decimal

+++ app/backends/webapp/src/main/scala/com/shocktrade/webapp/vm/dao/ClosePortfolioResponse.scala
@@ -12,4 +12,4 @@ import scala.scalajs.js
  */
 class ClosePortfolioResponse(val proceeds: js.UndefOr[PortfolioEquity],
                              val recommendation: js.UndefOr[AwardsRecommendation],
-                             val xp: js.UndefOr[Int]) extends js.Object
+                             val xp: js.UndefOr[Double]) extends js.Object
